Reorganize viz and lesson components in dir
