+++ algowiz-app/src/App.js
@@ -12,7 +12,7 @@ import {Container, Row, Col} from 'react-bootstrap';
 import Navbar from './components/NavBar';
 import HomePage from './components/HomePage';
 import QuizPage from './components/QuizPage';
-import LessonPage from './components/LessonPage';
+import LessonPage from './components/lessonComps/LessonPage';
 import Gutter from './components/Gutter';
 
 function App() {
